pyproject.toml
[project]
name = "aicommit"
version = "0.0.1"
authors = [
  { name="Abi Raja", email="abimanyuraja@gmail.com" },
]
description = "Generate AI powered commit messages"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    'async-timeout ~= 4.0',
    'black ~= 22.12.0',
    'certifi ~= 2022.12.7',
    'cffi ~= 1.15.1',
    'charset-normalizer ~= 2.1.1',
    'click ~= 8.1.3',
    'et-xmlfile ~= 1.1.0',
    'gitdb ~= 4.0.10',
    'GitPython ~= 3.1.29',
    'idna ~= 3.4',
    'importlib-metadata ~= 5.2.0',
    'jaraco.classes ~= 3.2.3',
    'keyring ~= 23.13.1',
    'langchain ~= 0.0.45',
    'Markdown ~= 3.4.1',
    'more-itertools ~= 9.0.0',
    'mypy-extensions ~= 0.4.3',
    'numpy ~= 1.24.0',
    'openai ~= 0.25.0',
    'openpyxl ~= 3.0.10',
    'pandas ~= 1.5.2',
    'pandas-stubs ~= 1.5.2.221213',
    'pathspec ~= 0.10.3',
    'platformdirs ~= 2.6.0',
    'prompt-toolkit ~= 1.0.14',
    'pycparser ~= 2.21',
    'pydantic ~= 1.10.2',
    'pygit2 ~= 1.11.1',
    'Pygments ~= 2.13.0',
    'PyInquirer ~= 1.0.3',
    'python-dateutil ~= 2.8.2',
    'python-dotenv ~= 0.21.0',
    'pytz ~= 2022.7',
    'PyYAML ~= 6.0',
    'redis ~= 4.4.0',
    'regex ~= 2022.10.31',
    'requests ~= 2.28.1',
    'six ~= 1.16.0',
    'smmap ~= 5.0.0',
    'SQLAlchemy ~= 1.4.45',
    'tomli ~= 2.0.1',
    'tqdm ~= 4.64.1',
    'types-pytz ~= 2022.7.0.0',
    'typing_extensions ~= 4.4.0',
    'urllib3 ~= 1.26.13',
    'wcwidth ~= 0.2.5',
    'zipp ~= 3.11.0'
]

[project.urls]
"Homepage" = "https://github.com/abi/autocommit"
"Bug Tracker" = "https://github.com/abi/autocommit/issues"

[project.scripts]
autocommit = "autocommit.commit:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
